Refactor sample objects
